refactor(streaming): unify streaming manager
